requirements.txt
anyio==4.4.0
arrow==1.3.0
asttokens==2.4.1
beautifulsoup4==4.12.3
binaryornot==0.4.4
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
dnspython==2.6.1
email_validator==2.1.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
flet==0.22.1
flet-core==0.22.1
flet-runtime==0.22.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
icecream==2.1.3
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
oauthlib==3.2.2
orjson==3.10.3
packaging==23.2
pillow==10.3.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
PyYAML==6.0.1
qrcode==7.4.2
repath==0.9.0
requests==2.32.2
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
text-unidecode==1.3
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.0
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.0
watchdog==4.0.1
watchfiles==0.22.0
websockets==12.0